weather-assistant/pyproject.toml
[project]
name = "weather-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muhammad Aneeque", email = "aneeq.khatri@celeritasdigital.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.8",
    "python-dotenv>=1.1.0",
]

[project.scripts]
weather-assistant = "weather_assistant:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
